move split file back
